split nym-nodes http handlers
