+++ nym-api/src/support/http/mod.rs
@@ -1,17 +1,16 @@
 // Copyright 2022-2023 - Nym Technologies SA <contact@nymtech.net>
 // SPDX-License-Identifier: GPL-3.0-only
 
+use nym_task::TaskManager;
+use tokio::task::JoinHandle;
+use tokio_util::sync::CancellationToken;
+
 pub(crate) mod helpers;
 pub(crate) mod openapi;
 pub(crate) mod router;
 pub(crate) mod state;
 
-use nym_task::TaskManager;
 pub(crate) use router::RouterBuilder;
-use tokio::task::JoinHandle;
-use tokio_util::sync::CancellationToken;
-
-use crate::unstable_routes;
 
 pub(crate) const TASK_MANAGER_TIMEOUT_S: u64 = 10;
