+++ nym-api/src/support/http/router.rs
@@ -11,7 +11,7 @@ use crate::nym_nodes::handlers::nym_node_routes;
 use crate::status;
 use crate::support::http::openapi::ApiDoc;
 use crate::support::http::state::AppState;
-use crate::support::http::unstable_routes::unstable_routes;
+use crate::unstable_routes::v1::unstable_routes_v1;
 use anyhow::anyhow;
 use axum::response::Redirect;
 use axum::routing::get;
@@ -64,7 +64,7 @@ impl RouterBuilder {
                     .nest("/api-status", status::handlers::api_status_routes())
                     .nest("/nym-nodes", nym_node_routes())
                     .nest("/ecash", ecash_routes())
-                    .nest("/unstable", unstable_routes()), // CORS layer needs to be "outside" of routes
+                    .nest("/unstable", unstable_routes_v1()), // CORS layer needs to be "outside" of routes
             );
 
         Self {
