+++ nym-api/src/support/http/state/mod.rs
@@ -16,8 +16,8 @@ use crate::support::http::state::chain_status::ChainStatusCache;
 use crate::support::http::state::force_refresh::ForcedRefresh;
 use crate::support::nyxd::Client;
 use crate::support::storage;
-use crate::unstable_routes::account::cache::AddressInfoCache;
-use crate::unstable_routes::models::NyxAccountDetails;
+use crate::unstable_routes::v1::account::cache::AddressInfoCache;
+use crate::unstable_routes::v1::account::models::NyxAccountDetails;
 use axum::extract::FromRef;
 use nym_api_requests::models::{GatewayBondAnnotated, MixNodeBondAnnotated, NodeAnnotation};
 use nym_mixnet_contract_common::NodeId;
