+++ nym-api/src/unstable_routes/helpers.rs
@@ -1,6 +1,2 @@
 // Copyright 2025 - Nym Technologies SA <contact@nymtech.net>
 // SPDX-License-Identifier: GPL-3.0-only
-
-pub(crate) mod helpers;
-pub(crate) mod v1;
-pub(crate) mod v2;
