+++ nym-api/src/unstable_routes/v1/nym_nodes/full_fat/mod.rs
@@ -1,9 +1,9 @@
-// Copyright 2024 - Nym Technologies SA <contact@nymtech.net>
+// Copyright 2025 - Nym Technologies SA <contact@nymtech.net>
 // SPDX-License-Identifier: GPL-3.0-only
 
 use crate::node_status_api::models::{AxumErrorResponse, AxumResult};
-use crate::nym_nodes::handlers::unstable::NodesParamsWithRole;
 use crate::support::http::state::AppState;
+use crate::unstable_routes::v1::nym_nodes::helpers::NodesParamsWithRole;
 use axum::extract::{Query, State};
 use nym_api_requests::nym_nodes::{CachedNodesResponse, FullFatNode};
 use nym_http_api_common::FormattedResponse;
@@ -15,11 +15,11 @@ use nym_http_api_common::FormattedResponse;
     path = "",
     context_path = "/v1/unstable/nym-nodes/full-fat",
     responses(
-    // (status = 200, body = CachedNodesResponse<FullFatNode>)
+        // (status = 200, body = CachedNodesResponse<FullFatNode>)
         (status = 501)
     )
 )]
-pub(super) async fn nodes_detailed(
+pub(crate) async fn nodes_detailed(
     _state: State<AppState>,
     _query_params: Query<NodesParamsWithRole>,
 ) -> AxumResult<FormattedResponse<CachedNodesResponse<FullFatNode>>> {
