+++ nym-api/src/unstable_routes/v1/nym_nodes/semi_skimmed/mod.rs
@@ -1,9 +1,9 @@
-// Copyright 2024 - Nym Technologies SA <contact@nymtech.net>
+// Copyright 2025 - Nym Technologies SA <contact@nymtech.net>
 // SPDX-License-Identifier: GPL-3.0-only
 
 use crate::node_status_api::models::{AxumErrorResponse, AxumResult};
-use crate::nym_nodes::handlers::unstable::NodesParamsWithRole;
 use crate::support::http::state::AppState;
+use crate::unstable_routes::v1::nym_nodes::helpers::NodesParamsWithRole;
 use axum::extract::{Query, State};
 use nym_api_requests::nym_nodes::{CachedNodesResponse, SemiSkimmedNode};
 use nym_http_api_common::FormattedResponse;
@@ -15,11 +15,11 @@ use nym_http_api_common::FormattedResponse;
     path = "",
     context_path = "/v1/unstable/nym-nodes/semi-skimmed",
     responses(
-    // (status = 200, body = CachedNodesResponse<SemiSkimmedNode>)
+        // (status = 200, body = CachedNodesResponse<SemiSkimmedNode>)
         (status = 501)
     )
 )]
-pub(super) async fn nodes_expanded(
+pub(crate) async fn nodes_expanded(
     _state: State<AppState>,
     _query_params: Query<NodesParamsWithRole>,
 ) -> AxumResult<FormattedResponse<CachedNodesResponse<SemiSkimmedNode>>> {
